# swagger.yaml
openapi: 3.0.0
info:
  title: LCMS API
  description: Tài liệu API cho hệ thống Quản lý Trung tâm Học tập (LCMS)
  version: 2.0.0
servers:
  - url: http://localhost:3003/api/v1
    description: Local Development Server
security:
  - bearerAuth: []
...
paths:
...
  /branch/create-branch:
    post:
      tags: [Branch]
      summary: Tạo cơ sở mới — SO (1.1)
      description: Chủ hệ thống chỉ cấp email và mật khẩu ban đầu. API tự tạo hồ sơ cơ sở tạm thời và tài khoản chủ cơ sở để người này đăng nhập rồi cấu hình tiếp.
      operationId: createBranch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBranchRequest'
      responses:
        '201':
          description: Tạo cơ sở thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Branch'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /branch/my-overview:
    get:
      tags: [Branch]
      summary: Tổng quan cơ sở hiện tại — BO/STAFF (1.2a)
      description: Lấy các thông tin tổng quan cho cơ sở gắn với tài khoản đang đăng nhập, dùng cho màn hình sau đăng nhập của chủ cơ sở hoặc nhân viên.
      operationId: getMyBranchOverview
      responses:
        '200':
          description: Tổng quan cơ sở hiện tại
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BranchOverview'
        '401': { $ref: '#/components/responses/UnauthorizedError' }
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /branch/{id}/overview:
    get:
      tags: [Branch]
      summary: Tổng quan cơ sở — SO/BO/STAFF (1.2b)
      operationId: getBranchOverview
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Tổng quan cơ sở
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BranchOverview'
        '403': { $ref: '#/components/responses/ForbiddenError' }
        '404': { $ref: '#/components/responses/NotFoundError' }
  /branch/{id}:
    get:
      tags: [Branch]
      summary: Chi tiết cơ sở (1.3)
      operationId: getBranchById
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Chi tiết cơ sở
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Branch'
        '404': { $ref: '#/components/responses/NotFoundError' }

    patch:
      tags: [Branch]
      summary: Cập nhật cơ sở — SO/BO (1.4)
      description: Chủ cơ sở chỉ được cập nhật cơ sở của mình với các trường hồ sơ và cấu hình mặc định bên dưới. Mã cơ sở, chủ sở hữu, trạng thái hoạt động và danh sách phòng học do các chức năng quản trị riêng quản lý.
      operationId: updateBranch
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateBranchRequest'
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Branch'
        '404': { $ref: '#/components/responses/NotFoundError' }
...
  /class/create-class:
    post:
      tags: [Class]
      summary: Tạo lớp học mới — SO/BO/STAFF (4.1)
      operationId: createClass
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClassRequest'
      responses:
        '201':
          description: Tạo lớp thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Class'
        '422': { $ref: '#/components/responses/ValidationError' }
...
  /class/{id}:
    get:
      tags: [Class]
      summary: Chi tiết lớp học (4.3)
      operationId: getClassById
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Chi tiết lớp học
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Class'
        '404': { $ref: '#/components/responses/NotFoundError' }

    patch:
      tags: [Class]
      summary: Cập nhật lớp học — SO/BO/STAFF (4.4)
      operationId: updateClass
      parameters:
        - $ref: '#/components/parameters/id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClassRequest'
      responses:
        '200':
          description: Cập nhật thành công
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Class'
        '404': { $ref: '#/components/responses/NotFoundError' }
  /class/{id}/close:
    patch:
      tags: [Class]
      summary: Đóng lớp học — SO/BO/STAFF (4.5)
      operationId: closeClass
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        '200':
          description: Đóng lớp thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404': { $ref: '#/components/responses/NotFoundError' }
  /class/{id}/students:
    get:
      tags: [Class]
      summary: Danh sách học sinh trong lớp (4.6)
      operationId: getClassStudents
      parameters:
        - $ref: '#/components/parameters/id'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          description: Lọc theo trạng thái — mặc định `active`
          schema:
            type: string
            enum: [active, left]
            default: active
      responses:
        '200':
          description: Danh sách học sinh trong lớp
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ClassStudentItem'
                      meta:
                        $ref: '#/components/schemas/PaginationMeta'
        '404': { $ref: '#/components/responses/NotFoundError' }

  # ── Enrollment (Module 5) ─────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    page:
      name: page
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string, example: "Thành công" }
        requestId: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
    ApiErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: "Lỗi" }
        code: { type: string, example: "NOT_FOUND" }
        requestId: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        details: { type: object }
    PaginationMeta:
      type: object
      properties:
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        total_pages: { type: integer }

    # ── Auth ──────────────────────────────────────────────────────
...
    Branch:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        code: { type: string }
        address: { type: string }
        phone: { type: string }
        email: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    BranchOverview:
      type: object
      properties:
        branch_id: { type: string }
        branch_code: { type: string }
        name: { type: string }
        owner:
          type: object
          nullable: true
          properties:
            _id: { type: string }
            full_name: { type: string }
            email: { type: string, format: email }
            phone: { type: string }
            is_active: { type: boolean }
        contact:
          type: object
          properties:
            email: { type: string, format: email }
            phone: { type: string }
            address: { type: string }
        status:
          type: object
          properties:
            is_active: { type: boolean }
            deleted_at: { type: string, format: date-time, nullable: true }
        settings:
          type: object
          properties:
            timezone: { type: string }
            default_fee_per_session: { type: number }
            default_session_slots:
              type: array
              items:
                type: object
            rooms:
              type: array
              items:
                type: object
        member_summary:
          type: object
          properties:
            total_members: { type: integer }
            total_accounts: { type: integer }
            by_status:
              type: object
              properties:
                total: { type: integer }
                active: { type: integer }
                inactive: { type: integer }
                deleted: { type: integer }
            by_role:
              type: object
              additionalProperties:
                type: object
                properties:
                  total: { type: integer }
                  active: { type: integer }
                  inactive: { type: integer }
                  deleted: { type: integer }
        class_summary:
          type: object
          properties:
            total: { type: integer }
            by_status:
              type: object
              additionalProperties: { type: integer }
        finance_summary:
          type: object
          properties:
            revenue_this_month: { type: number }
            invoices_total: { type: integer }
            invoices_by_status:
              type: object
              additionalProperties: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        as_of: { type: string, format: date-time }
    CreateBranchRequest:
      type: object
      required: [ownerEmail, password]
      properties:
        ownerEmail:
          type: string
          format: email
          example: "owner.branch@lcms.edu.vn"
        password:
          type: string
          format: password
          minLength: 8
          example: "Owner@123"
    UpdateBranchRequest:
      type: object
      properties:
        name: { type: string, maxLength: 150 }
        address: { type: string, maxLength: 500 }
        phone:
          type: string
          maxLength: 30
          pattern: '^[0-9+().\s-]*$'
        email: { type: string, format: email }
        timezone: { type: string, example: "Asia/Ho_Chi_Minh" }
        defaultFeePerSession:
          type: number
          minimum: 0
          nullable: true
        defaultSessionSlots:
          type: array
          maxItems: 20
          items:
            type: object
            required: [name, startTime, endTime]
            properties:
              name: { type: string, maxLength: 80 }
              startTime: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$' }
              endTime: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$' }

    # ── User ──────────────────────────────────────────────────────
...
    Class:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        code: { type: string }
        branchId: { type: string }
        teacherId: { type: string }
        subjectName: { type: string }
        status: { type: string, enum: [active, closed] }
        studentCount: { type: integer }
        maxStudents: { type: integer }
        schedule: { type: string }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        tuitionFee: { type: number }
        createdAt: { type: string, format: date-time }
    CreateClassRequest:
      type: object
      required: [name, teacher_id, subject_name, tuition_fee]
      properties:
        name: { type: string }
        teacher_id: { type: string }
        subject_name: { type: string }
        max_students: { type: integer }
        schedule: { type: string }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        tuition_fee: { type: number, minimum: 0 }
    UpdateClassRequest:
      type: object
      properties:
        name: { type: string }
        teacher_id: { type: string }
        subject_name: { type: string }
        max_students: { type: integer }
        schedule: { type: string }
        tuition_fee: { type: number, minimum: 0 }

    # ── Enrollment ────────────────────────────────────────────────
    ClassStudentItem:
      type: object
      description: Enrollment record với studentId được populate
      properties:
        _id: { type: string, description: Enrollment ID }
        enrolledAt: { type: string, format: date-time, description: Ngày tham gia }
        leftAt: { type: string, format: date-time, nullable: true }
        leftReason: { type: string, nullable: true }
        studentId:
          type: object
          description: Thông tin học sinh
          properties:
            _id: { type: string }
            fullName: { type: string, description: Họ tên học sinh }
            userCode: { type: string, description: Mã học sinh }
            phone: { type: string, nullable: true, description: Số điện thoại liên hệ }
            email: { type: string, format: email, description: Email liên hệ }
            dateOfBirth: { type: string, format: date, nullable: true }
...
  responses:
    UnauthorizedError:
      description: Chưa đăng nhập hoặc Token không hợp lệ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    ForbiddenError:
      description: Không có quyền truy cập
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFoundError:
      description: Không tìm thấy tài nguyên
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
    ValidationError:
      description: Dữ liệu gửi lên không đúng định dạng
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'